[in progress: 865] Live stream analytics grafikon, a default ertek csak nagyon picike legyen

+++ modules/Visitor/Live/Controller.php
@@ -718,7 +718,7 @@ class Controller extends \Visitor\Controller {
         $key = $fieldtokey[ $field ];
         $ret['data'][ $key ][] = array(
           'x' => $ts,
-          'y' => intval( $v ),
+          'y' => floatval( $v ),
         );
 
       }
